[32기 정덕우]
