+++ src/components/ItemContainer/ItemContainer.js
@@ -1,33 +1,34 @@
-import React, { useState, useEffect } from "react";
-import { useLocation, useNavigate, useParams } from "react-router-dom";
 import Items from "../Items/Items";
-import { config } from "../../config";
 import "./ItemContainer.scss";
 
 const LIMIT = 4;
 
-const ItemContainer = ({ title, name, itemLists, setItemLists }) => {
-  const [page, setPage] = useState(1);
-  const navigate = useNavigate();
-  const location = useLocation();
-  const params = useParams();
+const ItemContainer = ({ title, name, itemLists, updateOffset }) => {
+  // const [page, setPage] = useState(1);
+  // const navigate = useNavigate();
+  // const location = useLocation();
+  // const params = useParams();
 
-  useEffect(() => {
-    fetch(
-      `${config.category}/${params.id}${
-        location.search || `?limit=${LIMIT}&offset=0`
-      }`
-    )
-      .then(res => res.json())
-      .then(res => setItemLists(res.products));
-  }, [location.search]);
+  // useEffect(() => {
+  //   fetch(
+  //     `${config.category}/${params.id}${
+  //       location.search || `?limit=${LIMIT}&offset=0`
+  //     }`
+  //   )
+  //     .then(res => res.json())
+  //     .then(res => {
+  //       setItemLists(prev => prev.concat(res.products));
+  //     });
+  // }, [location.search]);
 
-  const updateOffset = e => {
-    const offset = page * LIMIT;
-    const queryString = `?limit=${LIMIT}&offset=${offset}`;
-    setPage(page + 1);
-    navigate(`${queryString}`);
-  };
+  // const updateOffset = e => {
+  //   const offset = page * LIMIT;
+  //   const queryString = `?limit=${LIMIT}&offset=${offset}`;
+  //   setPage(page + 1);
+  //   navigate(`${queryString}`);
+  // };
+
+  console.log(itemLists);
 
   return (
     <div className="itemContainer">
