+++ src/pages/Category/Category.js
@@ -1,23 +1,48 @@
 import React, { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
+import { useParams, useLocation, useNavigate } from "react-router-dom";
 import ItemContainer from "../../components/ItemContainer/ItemContainer";
 import "./Category.scss";
+import { config } from "../../config";
+
+const LIMIT = 4;
+let PAGE = 1;
 
 const Category = () => {
   const [itemLists, setItemLists] = useState([]);
   const [categoryInfo, setCategoryInfo] = useState([]);
+  const navigate = useNavigate();
+  const location = useLocation();
   const params = useParams();
 
   useEffect(() => {
     fetch(
-      `http://10.58.1.230:8000/products/category/${params.id}?sort_method=0&offset=0&limit=4`
+      `${config.category}/${params.id}${
+        location.search || `?limit=${LIMIT}&offset=0`
+      }`
     )
       .then(res => res.json())
       .then(res => {
-        setItemLists(res.products);
-        setCategoryInfo(res.category);
+        const copyArray = [...res.products];
+        setItemLists(itemLists.concat(copyArray));
       });
-  }, [params.id]);
+  }, [PAGE]);
+
+  const updateOffset = e => {
+    const offset = PAGE * LIMIT;
+    const queryString = `?limit=${LIMIT}&offset=${offset}`;
+    PAGE += 1;
+    navigate(`${queryString}`);
+  };
+  // useEffect(() => {
+  //   fetch(
+  //     `http://10.58.1.230:8000/products/category/${params.id}?sort_method=0&offset=0&limit=4`
+  //   )
+  //     .then(res => res.json())
+  //     .then(res => {
+  //       setItemLists(res.products);
+  //       setCategoryInfo(res.category);
+  //     });
+  // }, [params.id]);
 
   return (
     <>
@@ -42,7 +67,7 @@ const Category = () => {
         name="itemsInCategory"
         title="요즘 잘 나가요"
         itemLists={itemLists}
-        setItemLists={setItemLists}
+        updateOffset={updateOffset}
       />
     </>
   );
